revert most include namespace cleanup
C include files which are internal to a project should be include via double-quotes, e.g.

 #include "internal.h"

while those for broader use should use angle brackets:

 #include <external.h>

the previous effort to inflict conformity on the code base did not
take into account the large number of include files which were
installed.  Some are known not public, and are no longer installed,
but an overall assessment of which need to be installed was not
performed.

Therefore, all installed include files now use the angle-bracket notation
to include each other.